requirements.txt
fastapi
uvicorn[standard]
python-dotenv
jinja2
requests
